Updated users api; Makes sure target user is not an employee or manager before deleting getting deleted

+++ app/api/users/delete/route.ts
@@ -1,5 +1,7 @@
 import UserModel from "@/models/userModel";
 import { connectDB } from "@/lib/database";
+import Employee from "@/models/employeeModel";
+import Manager from "@/models/managerModel";
 
 export const DELETE = async (request: Request) => {
   const { id } = await request.json();
@@ -9,11 +11,29 @@ export const DELETE = async (request: Request) => {
   }
 
   await connectDB();
-  const user = await UserModel.findById(id).exec();
+  const [user, employee] = await Promise.all([
+    UserModel.findById(id).lean().exec(),
+    Employee.findOne({ user: id }).lean().exec(),
+  ]);
+
+  const manager = await Manager.findOne({ employee: employee?._id })
+    .lean()
+    .exec();
 
   if (!user) {
     return new Response("User not found", { status: 400 });
   }
+
+  if (employee)
+    return new Response("Make sure to remove user as an employee first ", {
+      status: 400,
+    });
+
+  if (manager)
+    return new Response(
+      "Make sure to demote user from manager status and remove user as an employee ",
+      { status: 400 }
+    );
 
   await user.deleteOne();
